Use juel for reflection (unfinished)
