+++ src/main/java/com/hubspot/jinjava/el/JinjavaInterpreterResolver.java
@@ -12,13 +12,15 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import javax.el.ELContext;
+import javax.el.PropertyNotFoundException;
 
 import org.apache.commons.lang3.LocaleUtils;
 import org.apache.commons.lang3.StringUtils;
 
-import com.google.common.collect.Lists;
 import com.hubspot.jinjava.el.ext.AbstractCallableMethod;
 import com.hubspot.jinjava.el.ext.ExtendedParser;
 import com.hubspot.jinjava.interpret.JinjavaInterpreter;
@@ -32,7 +34,6 @@ import com.hubspot.jinjava.objects.date.FormattedDate;
 import com.hubspot.jinjava.objects.date.PyishDate;
 import com.hubspot.jinjava.objects.date.StrftimeFormatter;
 import com.hubspot.jinjava.util.JinjavaPropertyNotResolvedException;
-
 import de.odysseus.el.util.SimpleResolver;
 
 public class JinjavaInterpreterResolver extends SimpleResolver {
@@ -77,8 +78,8 @@ public class JinjavaInterpreterResolver extends SimpleResolver {
         value = interpreter.retraceVariable((String) prop, interpreter.getLineNumber());
       } else {
         try {
-          value = interpreter.resolve(base, Lists.newArrayList(property));
-        } catch (JinjavaPropertyNotResolvedException e) {
+          value = super.getValue(context, base, transformName(property));
+        } catch (PropertyNotFoundException | JinjavaPropertyNotResolvedException e) {
           if (errOnUnknownProp) {
             interpreter.addError(TemplateError.fromUnknownProperty(base, property, interpreter.getLineNumber()));
           }
@@ -153,6 +154,23 @@ public class JinjavaInterpreterResolver extends SimpleResolver {
     }
 
     return Locale.US;
+  }
+
+  // snake case stuff
+
+  private static final Pattern SNAKE_CASE = Pattern.compile("_([^_]?)");
+
+  private String transformName(String name) {
+    Matcher m = SNAKE_CASE.matcher(name);
+
+    StringBuffer result = new StringBuffer(name.length());
+    while (m.find()) {
+      String replacement = m.group(1).toUpperCase();
+      m.appendReplacement(result, replacement);
+    }
+    m.appendTail(result);
+
+    return result.toString();
   }
 
 }
